requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
--index-url http://edx.devstack.devpi:3141/root/pypi/+simple/
--extra-index-url https://pypi.python.org/simple
--trusted-host edx.devstack.devpi

asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
astroid==3.2.4 ; python_version < "3.9"
    # via
    #   -c requirements/constraints.txt
    #   pylint
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
backports-zoneinfo==0.2.1 ; python_version < "3.9"
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   django
cachetools==5.5.0
    # via tox
chardet==5.2.0
    # via tox
colorama==0.4.6
    # via tox
coverage==7.6.1
    # via -r requirements/test.in
dill==0.3.9
    # via pylint
distlib==0.3.9
    # via virtualenv
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
exceptiongroup==1.2.2
    # via pytest
filelock==3.16.1
    # via
    #   tox
    #   virtualenv
importlib-resources==6.4.5
    # via
    #   jsonschema
    #   jsonschema-specifications
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via
    #   -r requirements/test.in
    #   pylint
jsonschema==4.23.0
    # via -r requirements/test.in
jsonschema-specifications==2023.12.1
    # via jsonschema
mccabe==0.7.0
    # via pylint
packaging==24.2
    # via
    #   pyproject-api
    #   pytest
    #   tox
pkgutil-resolve-name==1.3.10
    # via jsonschema
platformdirs==4.3.6
    # via
    #   pylint
    #   tox
    #   virtualenv
pluggy==1.5.0
    # via
    #   pytest
    #   tox
pycodestyle==2.12.1
    # via -r requirements/test.in
pydocstyle==6.3.0
    # via -r requirements/test.in
pylint==3.2.7
    # via
    #   -r requirements/test.in
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.5
    # via -r requirements/test.in
pylint-plugin-utils==0.8.2
    # via pylint-django
pyproject-api==1.8.0
    # via tox
pytest==8.3.3
    # via pytest-django
pytest-django==4.9.0
    # via -r requirements/test.in
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.20.1
    # via
    #   jsonschema
    #   referencing
snowballstemmer==2.2.0
    # via pydocstyle
sqlparse==0.5.2
    # via
    #   -r requirements/base.txt
    #   django
tomli==2.1.0
    # via
    #   pylint
    #   pyproject-api
    #   pytest
    #   tox
tomlkit==0.13.2
    # via pylint
tox==4.23.2
    # via -r requirements/test.in
typing-extensions==4.12.2
    # via
    #   -r requirements/base.txt
    #   asgiref
    #   astroid
    #   pylint
    #   tox
virtualenv==20.27.1
    # via tox
zipp==3.20.2
    # via importlib-resources
